requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Caching==2.3.1
redis==6.4.0
gunicorn==21.2.0
Pillow==10.0.1
psycopg2-binary==2.9.7
